Noticed some mistakes I made in the initial conversion from their JSON example to C# classes, also that they define a bunch of items as floats when they actually appear to be decimal? The note in their docs says they convert it to float on their side...

+++ TaxService/Models/TaxJar/CalculateTax/Request/TaxJarCalculateTax_Model.cs
@@ -38,10 +38,10 @@ namespace TaxService.Models.TaxJar
         public string ToStreet { get; set; }
 
         [JsonProperty("amount")]
-        public float? Amount { get; set; }
+        public decimal? Amount { get; set; }
 
         [JsonProperty("shipping")]
-        public float Shipping { get; set; }
+        public decimal Shipping { get; set; }
 
         [JsonProperty("customer_id")]
         public string CustomerID { get; set; }
